警告の抑止。 buffはたかだか20バイトであるため、intキャストで問題ない。
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@4499 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ TTXSamples/TTXttyrec/TTXttyrec.c
@@ -231,7 +231,7 @@ static int PASCAL FAR TTXProcessCommand(HWND hWin, WORD cmd) {
 	  if (pvar->rec_stsize) {
 	    _snprintf_s(buff, sizeof(buff), _TRUNCATE, "\033[8;%d;%dt",
 	                pvar->ts->TerminalHeight, pvar->ts->TerminalWidth);
-	    WriteData(pvar->fh, buff, strlen(buff));
+	    WriteData(pvar->fh, buff, (int)strlen(buff));
 	  }
 	}
       }
